+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision History
 
+## 1.1 (unreleased)
+
+- Added coloring to the command-line output.
+- Added preliminary Windows support.
+
 ## 1.0.2 (2016/07/28)
 
 - Moved documentation to http://gitman.readthedocs.io/.